refactor: (WIP) Separate UI from business rule
