+++ src/modules/Product/composables/Product/useFormPresentation.ts
@@ -1,0 +1,10 @@
+//import { onMounted, reactive, ref, watch, computed } from 'vue'
+// import type { Presentation } from "../../types/Presentation";
+
+export default (presentation: Presentation) => {
+  
+
+  return {
+     
+  }
+}
